Break up opds1.py into multiple files
