+++ src/palace/manager/integration/license/opds/opds1/importer.py
@@ -2,14 +2,10 @@ from __future__ import annotations
 
 import logging
 import traceback
-import urllib
-from abc import ABC, abstractmethod
-from collections import defaultdict
-from collections.abc import Callable, Generator, Iterable, Mapping, Sequence
+from collections.abc import Generator, Sequence
 from datetime import datetime
-from enum import Enum
 from io import BytesIO
-from typing import TYPE_CHECKING, Any, Generic, TypeVar, overload
+from typing import Any, overload
 from urllib.parse import urljoin, urlparse
 from xml.etree.ElementTree import Element
 
@@ -18,633 +14,33 @@ import feedparser
 from feedparser import FeedParserDict
 from flask_babel import lazy_gettext as _
 from lxml import etree
-from requests import Response
-from sqlalchemy.orm.session import Session
-from typing_extensions import Unpack
+from sqlalchemy.orm import Session
 
-from palace.manager.api.circulation.base import BaseCirculationAPI
-from palace.manager.api.circulation.data import HoldInfo, LoanInfo
-from palace.manager.api.circulation.exceptions import (
-    CurrentlyAvailable,
-    FormatNotAvailable,
-    NotOnHold,
-)
-from palace.manager.api.circulation.fulfillment import RedirectFulfillment
-from palace.manager.api.circulation.settings import BaseCirculationApiSettings
-from palace.manager.core.classifier import Classifier
 from palace.manager.core.coverage import CoverageFailure
-from palace.manager.core.monitor import CollectionMonitor, TimestampData
 from palace.manager.data_layer.bibliographic import BibliographicData
 from palace.manager.data_layer.circulation import CirculationData
 from palace.manager.data_layer.contributor import ContributorData
 from palace.manager.data_layer.identifier import IdentifierData
 from palace.manager.data_layer.link import LinkData
 from palace.manager.data_layer.measurement import MeasurementData
-from palace.manager.data_layer.policy.replacement import ReplacementPolicy
 from palace.manager.data_layer.subject import SubjectData
-from palace.manager.integration.base import integration_settings_load
-from palace.manager.integration.license.opds.settings.format_priority import (
-    FormatPriorities,
-    FormatPrioritiesSettings,
+from palace.manager.integration.license.opds.base.importer import BaseOPDSImporter
+from palace.manager.integration.license.opds.opds1.api import OPDSAPI
+from palace.manager.integration.license.opds.opds1.settings import (
+    IdentifierSource,
+    OPDSImporterSettings,
 )
-from palace.manager.integration.license.opds.settings.wayfless import (
-    SAMLWAYFlessConstants,
-    SAMLWAYFlessFulfillmentError,
-    SAMLWAYFlessSetttings,
-)
-from palace.manager.integration.license.settings.connection import ConnectionSetting
-from palace.manager.integration.patron_auth.saml.credential import SAMLCredentialManager
-from palace.manager.integration.settings import (
-    BaseSettings,
-    ConfigurationFormItem,
-    ConfigurationFormItemType,
-    FormField,
-)
+from palace.manager.integration.license.opds.opds1.xml_parser import OPDSXMLParser
 from palace.manager.sqlalchemy.model.classification import Subject
 from palace.manager.sqlalchemy.model.collection import Collection
-from palace.manager.sqlalchemy.model.coverage import CoverageRecord
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.identifier import Identifier
-from palace.manager.sqlalchemy.model.licensing import (
-    LicensePool,
-    LicensePoolDeliveryMechanism,
-    RightsStatus,
-)
+from palace.manager.sqlalchemy.model.licensing import RightsStatus
 from palace.manager.sqlalchemy.model.measurement import Measurement
-from palace.manager.sqlalchemy.model.patron import Patron
 from palace.manager.sqlalchemy.model.resource import Hyperlink
-from palace.manager.sqlalchemy.util import get_one
-from palace.manager.util import base64
-from palace.manager.util.datetime_helpers import datetime_utc, to_utc, utc_now
-from palace.manager.util.http import HTTP, BadResponseException
-from palace.manager.util.log import LoggerMixin
+from palace.manager.util.datetime_helpers import datetime_utc, utc_now
 from palace.manager.util.opds_writer import OPDSFeed, OPDSMessage
-from palace.manager.util.pydantic import HttpUrl
-from palace.manager.util.xmlparser import XMLParser
-
-if TYPE_CHECKING:
-    from palace.manager.sqlalchemy.model.work import Work
-
-
-class OPDSXMLParser(XMLParser):
-    NAMESPACES = {
-        "simplified": "http://librarysimplified.org/terms/",
-        "app": "http://www.w3.org/2007/app",
-        "dcterms": "http://purl.org/dc/terms/",
-        "dc": "http://purl.org/dc/elements/1.1/",
-        "opds": "http://opds-spec.org/2010/catalog",
-        "schema": "http://schema.org/",
-        "atom": "http://www.w3.org/2005/Atom",
-        "drm": "http://librarysimplified.org/terms/drm",
-        "palace": "http://palaceproject.io/terms",
-    }
-
-
-class IdentifierSource(Enum):
-    ID = "id"
-    DCTERMS_IDENTIFIER = "first_dcterms_identifier"
-
-
-class OPDSImporterSettings(
-    ConnectionSetting,
-    SAMLWAYFlessSetttings,
-    FormatPrioritiesSettings,
-    BaseCirculationApiSettings,
-):
-    external_account_id: HttpUrl = FormField(
-        form=ConfigurationFormItem(
-            label=_("URL"),
-            required=True,
-        )
-    )
-
-    data_source: str = FormField(
-        form=ConfigurationFormItem(label=_("Data source name"), required=True)
-    )
-
-    include_in_inventory_report: bool = FormField(
-        True,
-        form=ConfigurationFormItem(
-            label=_("Include in inventory report?"),
-            type=ConfigurationFormItemType.SELECT,
-            options={
-                True: "(Default) Yes",
-                False: "No",
-            },
-        ),
-    )
-
-    default_audience: str | None = FormField(
-        None,
-        form=ConfigurationFormItem(
-            label=_("Default audience"),
-            description=_(
-                "If the vendor does not specify the target audience for their books, "
-                "assume the books have this target audience."
-            ),
-            type=ConfigurationFormItemType.SELECT,
-            options={
-                **{None: _("No default audience")},
-                **{audience: audience for audience in sorted(Classifier.AUDIENCES)},
-            },
-            required=False,
-        ),
-    )
-
-    username: str | None = FormField(
-        None,
-        form=ConfigurationFormItem(
-            label=_("Username"),
-            description=_(
-                "If HTTP Basic authentication is required to access the OPDS feed (it usually isn't), enter the username here."
-            ),
-            weight=-1,
-        ),
-    )
-
-    password: str | None = FormField(
-        None,
-        form=ConfigurationFormItem(
-            label=_("Password"),
-            description=_(
-                "If HTTP Basic authentication is required to access the OPDS feed (it usually isn't), enter the password here."
-            ),
-            weight=-1,
-        ),
-    )
-
-    custom_accept_header: str = FormField(
-        default=",".join(
-            [
-                OPDSFeed.ACQUISITION_FEED_TYPE,
-                "application/atom+xml;q=0.9",
-                "application/xml;q=0.8",
-                "*/*;q=0.1",
-            ]
-        ),
-        form=ConfigurationFormItem(
-            label=_("Custom accept header"),
-            required=False,
-            description=_(
-                "Some servers expect an accept header to decide which file to send. You can use */* if the server doesn't expect anything."
-            ),
-            weight=-1,
-        ),
-    )
-
-    primary_identifier_source: IdentifierSource = FormField(
-        IdentifierSource.ID,
-        form=ConfigurationFormItem(
-            label=_("Identifer"),
-            required=False,
-            description=_("Which book identifier to use as ID."),
-            type=ConfigurationFormItemType.SELECT,
-            options={
-                IdentifierSource.ID: "(Default) Use <id>",
-                IdentifierSource.DCTERMS_IDENTIFIER: "Use <dcterms:identifier> first, if not exist use <id>",
-            },
-        ),
-    )
-
-
-class OPDSImporterLibrarySettings(BaseSettings):
-    pass
-
-
-class BaseOPDSAPI(
-    BaseCirculationAPI[OPDSImporterSettings, OPDSImporterLibrarySettings], ABC
-):
-    def __init__(self, _db: Session, collection: Collection):
-        super().__init__(_db, collection)
-        self.saml_wayfless_url_template = self.settings.saml_wayfless_url_template
-        self.saml_credential_manager = SAMLCredentialManager()
-        self._format_priorities = FormatPriorities(
-            self.settings.prioritized_drm_schemes,
-            self.settings.prioritized_content_types,
-            self.settings.deprioritize_lcp_non_epubs,
-        )
-
-    @property
-    def data_source(self) -> DataSource:
-        return DataSource.lookup(self._db, self.settings.data_source, autocreate=True)
-
-    def checkin(self, patron: Patron, pin: str, licensepool: LicensePool) -> None:
-        # All the CM side accounting for this loan is handled by CirculationAPI
-        # since we don't have any remote API we need to call this method is
-        # just a no-op.
-        pass
-
-    def release_hold(self, patron: Patron, pin: str, licensepool: LicensePool) -> None:
-        # Since there is no such thing as a hold, there is no such
-        # thing as releasing a hold.
-        raise NotOnHold()
-
-    def place_hold(
-        self,
-        patron: Patron,
-        pin: str | None,
-        licensepool: LicensePool,
-        notification_email_address: str | None,
-    ) -> HoldInfo:
-        # Because all OPDS content is assumed to be simultaneously
-        # available to all patrons, there is no such thing as a hold.
-        raise CurrentlyAvailable()
-
-    def update_availability(self, licensepool: LicensePool) -> None:
-        # We already know all the availability information we're going
-        # to know, so we don't need to do anything.
-        pass
-
-    def fulfill_saml_wayfless(
-        self, template: str, patron: Patron, fulfillment: RedirectFulfillment
-    ) -> RedirectFulfillment:
-        self.log.debug(f"WAYFless acquisition link template: {template}")
-
-        db = Session.object_session(patron)
-        saml_credential = self.saml_credential_manager.lookup_saml_token_by_patron(
-            db, patron
-        )
-
-        self.log.debug(f"SAML credentials: {saml_credential}")
-
-        if not saml_credential:
-            raise SAMLWAYFlessFulfillmentError(
-                f"There are no existing SAML credentials for patron {patron}"
-            )
-
-        saml_subject = self.saml_credential_manager.extract_saml_token(saml_credential)
-
-        self.log.debug(f"SAML subject: {saml_subject}")
-
-        if not saml_subject.idp:
-            raise SAMLWAYFlessFulfillmentError(
-                f"SAML subject {saml_subject} does not contain an IdP's entityID"
-            )
-
-        acquisition_link = template.replace(
-            SAMLWAYFlessConstants.IDP_PLACEHOLDER,
-            urllib.parse.quote(saml_subject.idp, safe=""),
-        )
-
-        acquisition_link = acquisition_link.replace(
-            SAMLWAYFlessConstants.ACQUISITION_LINK_PLACEHOLDER,
-            urllib.parse.quote(fulfillment.content_link, safe=""),
-        )
-
-        self.log.debug(
-            f"Old acquisition link {fulfillment.content_link} has been transformed to {acquisition_link}"
-        )
-
-        fulfillment.content_link = acquisition_link
-        return fulfillment
-
-    def fulfill(
-        self,
-        patron: Patron,
-        pin: str,
-        licensepool: LicensePool,
-        delivery_mechanism: LicensePoolDeliveryMechanism,
-        **kwargs: Unpack[BaseCirculationAPI.FulfillKwargs],
-    ) -> RedirectFulfillment:
-        requested_mechanism = delivery_mechanism.delivery_mechanism
-        rep = None
-        for lpdm in licensepool.available_delivery_mechanisms:
-            if (
-                lpdm.resource is None
-                or lpdm.resource.representation is None
-                or lpdm.resource.representation.public_url is None
-            ):
-                # This LicensePoolDeliveryMechanism can't actually
-                # be used for fulfillment.
-                continue
-            if lpdm.delivery_mechanism == requested_mechanism:
-                # We found it! This is how the patron wants
-                # the book to be delivered.
-                rep = lpdm.resource.representation
-                break
-
-        if not rep:
-            # There is just no way to fulfill this loan the way the
-            # patron wants.
-            raise FormatNotAvailable()
-
-        content_link = rep.public_url
-        media_type = rep.media_type
-
-        fulfillment_strategy = RedirectFulfillment(
-            content_link=content_link,
-            content_type=media_type,
-        )
-
-        if self.saml_wayfless_url_template:
-            fulfillment_strategy = self.fulfill_saml_wayfless(
-                self.saml_wayfless_url_template, patron, fulfillment_strategy
-            )
-
-        return fulfillment_strategy
-
-    def checkout(
-        self,
-        patron: Patron,
-        pin: str | None,
-        licensepool: LicensePool,
-        delivery_mechanism: LicensePoolDeliveryMechanism | None,
-    ) -> LoanInfo:
-        return LoanInfo.from_license_pool(licensepool, end_date=None)
-
-    def can_fulfill_without_loan(
-        self,
-        patron: Patron | None,
-        pool: LicensePool,
-        lpdm: LicensePoolDeliveryMechanism,
-    ) -> bool:
-        return True
-
-    def sort_delivery_mechanisms(
-        self, lpdms: list[LicensePoolDeliveryMechanism]
-    ) -> list[LicensePoolDeliveryMechanism]:
-        return self._format_priorities.prioritize_mechanisms(lpdms)
-
-
-SettingsType = TypeVar("SettingsType", bound=OPDSImporterSettings, covariant=True)
-
-
-class BaseOPDSImporter(
-    Generic[SettingsType],
-    LoggerMixin,
-    ABC,
-):
-    def __init__(
-        self,
-        _db: Session,
-        collection: Collection,
-        data_source_name: str | None,
-    ):
-        self._db = _db
-        if collection.id is None:
-            raise ValueError(
-                f"Unable to create importer for Collection with id = None. Collection: {collection.name}."
-            )
-        self._collection_id = collection.id
-        self._integration_configuration_id = collection.integration_configuration_id
-        if data_source_name is None:
-            # Use the Collection data_source for OPDS import.
-            data_source = self.collection.data_source
-            if data_source:
-                data_source_name = data_source.name
-            else:
-                raise ValueError(
-                    "Cannot perform an OPDS import on a Collection that has no associated DataSource!"
-                )
-        self.data_source_name = data_source_name
-        self.settings = integration_settings_load(
-            self.settings_class(), collection.integration_configuration
-        )
-
-    @classmethod
-    @abstractmethod
-    def settings_class(cls) -> type[SettingsType]: ...
-
-    @abstractmethod
-    def extract_feed_data(
-        self, feed: str | bytes, feed_url: str | None = None
-    ) -> tuple[dict[str, BibliographicData], dict[str, list[CoverageFailure]]]: ...
-
-    @abstractmethod
-    def extract_last_update_dates(
-        self, feed: str | bytes | FeedParserDict
-    ) -> list[tuple[str | None, datetime | None]]: ...
-
-    @abstractmethod
-    def extract_next_links(self, feed: str | bytes) -> list[str]: ...
-
-    @overload
-    def parse_identifier(self, identifier: str) -> Identifier: ...
-
-    @overload
-    def parse_identifier(self, identifier: str | None) -> Identifier | None: ...
-
-    def parse_identifier(self, identifier: str | None) -> Identifier | None:
-        """Parse the identifier and return an Identifier object representing it.
-
-        :param identifier: String containing the identifier
-
-        :return: Identifier object
-        """
-        parsed_identifier = None
-
-        try:
-            result = Identifier.parse_urn(self._db, identifier)
-            if result is not None:
-                parsed_identifier, _ = result
-        except Exception as e:
-            self.log.exception(
-                f"An unexpected exception occurred during parsing identifier '{identifier}': {e}"
-            )
-
-        return parsed_identifier
-
-    @property
-    def data_source(self) -> DataSource:
-        """Look up or create a DataSource object representing the
-        source of this OPDS feed.
-        """
-        offers_licenses = self.collection is not None
-        return DataSource.lookup(
-            self._db,
-            self.data_source_name,
-            autocreate=True,
-            offers_licenses=offers_licenses,
-        )
-
-    @property
-    def collection(self) -> Collection:
-        collection = Collection.by_id(self._db, self._collection_id)
-        if collection is None:
-            raise ValueError("Unable to load collection.")
-        return collection
-
-    def import_edition_from_bibliographic(
-        self, bibliographic: BibliographicData
-    ) -> Edition:
-        """For the passed-in BibliographicData object, see if can find or create an Edition
-        in the database. Also create a LicensePool if the BibliographicData has
-        CirculationData in it.
-        """
-        # Locate or create an Edition for this book.
-        edition, is_new_edition = bibliographic.edition(self._db)
-
-        policy = ReplacementPolicy(
-            subjects=True,
-            links=True,
-            contributions=True,
-            rights=True,
-            link_content=True,
-            formats=True,
-            even_if_not_apparently_updated=True,
-        )
-        bibliographic.apply(
-            self._db,
-            edition=edition,
-            collection=self.collection,
-            replace=policy,
-        )
-
-        return edition
-
-    def update_work_for_edition(
-        self,
-        edition: Edition,
-        is_open_access: bool = True,
-    ) -> tuple[LicensePool | None, Work | None]:
-        """If possible, ensure that there is a presentation-ready Work for the
-        given edition's primary identifier.
-
-        :param edition: The edition whose license pool and work we're interested in.
-        :param is_open_access: Whether this is an open access edition.
-        :return: 2-Tuple of license pool (optional) and work (optional) for edition.
-        """
-
-        work = None
-
-        # Looks up a license pool for the primary identifier associated with
-        # the given edition. If this is not an open access book, then the
-        # collection is also used as criteria for the lookup. Open access
-        # books don't require a collection match, according to this explanation
-        # from prior work:
-        #   Find a LicensePool for the primary identifier. Any LicensePool will
-        #   do--the collection doesn't have to match, since all
-        #   LicensePools for a given identifier have the same Work.
-        #
-        # If we have CirculationData, a pool was created when we
-        # imported the edition. If there was already a pool from a
-        # different data source or a different collection, that's fine
-        # too.
-        collection_criteria = {} if is_open_access else {"collection": self.collection}
-        pool = get_one(
-            self._db,
-            LicensePool,
-            identifier=edition.primary_identifier,
-            on_multiple="interchangeable",
-            **collection_criteria,
-        )
-
-        if pool:
-            if not pool.work or not pool.work.presentation_ready:
-                # There is no presentation-ready Work for this
-                # LicensePool. Try to create one.
-                work, ignore = pool.calculate_work()
-            else:
-                # There is a presentation-ready Work for this LicensePool.
-                # Use it.
-                work = pool.work
-
-        # If a presentation-ready Work already exists, there's no
-        # rush. We might have new BibliographicData that will change the Work's
-        # presentation, but when we called BibliographicData.apply() the work
-        # was set up to have its presentation recalculated in the
-        # background, and that's good enough.
-        return pool, work
-
-    def import_from_feed(self, feed: str | bytes, feed_url: str | None = None) -> tuple[
-        list[Edition],
-        list[LicensePool],
-        list[Work],
-        dict[str, list[CoverageFailure]],
-    ]:
-        # Keep track of editions that were imported. Pools and works
-        # for those editions may be looked up or created.
-        imported_editions = {}
-        pools = {}
-        works = {}
-
-        # If parsing the overall feed throws an exception, we should address that before
-        # moving on. Let the exception propagate.
-        bibliographic_objs, extracted_failures = self.extract_feed_data(feed, feed_url)
-        failures = defaultdict(list, extracted_failures)
-        # make editions.  if have problem, make sure associated pool and work aren't created.
-        for key, bibliographic in bibliographic_objs.items():
-            # key is identifier.urn here
-
-            # If there's a status message about this item, don't try to import it.
-            if key in list(failures.keys()):
-                continue
-
-            try:
-                # Create an edition. This will also create a pool if there's circulation data.
-                edition = self.import_edition_from_bibliographic(bibliographic)
-                if edition:
-                    imported_editions[key] = edition
-            except Exception as e:
-                # Rather than scratch the whole import, treat this as a failure that only applies
-                # to this item.
-                self.log.error("Error importing an OPDS item", exc_info=e)
-                data_source = self.data_source
-                identifier = bibliographic.load_primary_identifier(self._db)
-                failure = CoverageFailure(
-                    identifier,
-                    traceback.format_exc(),
-                    data_source=data_source,
-                    transient=False,
-                    collection=self.collection,
-                )
-                failures[key].append(failure)
-                # clean up any edition might have created
-                if key in imported_editions:
-                    del imported_editions[key]
-                # Move on to the next item, don't create a work.
-                continue
-
-            try:
-                pool, work = self.update_work_for_edition(edition)
-                if pool:
-                    pools[key] = pool
-                if work:
-                    works[key] = work
-            except Exception as e:
-                collection_name = self.collection.name if self.collection else "None"
-                logging.warning(
-                    f"Non-fatal exception: Failed to import item - import will continue: "
-                    f"identifier={key}; collection={collection_name}/{self._collection_id}; "
-                    f"data_source={self.data_source}; exception={e}",
-                    stack_info=True,
-                )
-                identifier, ignore = Identifier.parse_urn(self._db, key)
-                data_source = self.data_source
-                failure = CoverageFailure(
-                    identifier,
-                    traceback.format_exc(),
-                    data_source=data_source,
-                    transient=False,
-                    collection=self.collection,
-                )
-                failures[key].append(failure)
-
-        return (
-            list(imported_editions.values()),
-            list(pools.values()),
-            list(works.values()),
-            failures,
-        )
-
-
-class OPDSAPI(BaseOPDSAPI):
-    @classmethod
-    def settings_class(cls) -> type[OPDSImporterSettings]:
-        return OPDSImporterSettings
-
-    @classmethod
-    def library_settings_class(cls) -> type[OPDSImporterLibrarySettings]:
-        return OPDSImporterLibrarySettings
-
-    @classmethod
-    def description(cls) -> str:
-        return "Import books from a publicly-accessible OPDS feed."
-
-    @classmethod
-    def label(cls) -> str:
-        return "OPDS Import"
 
 
 class OPDSImporter(BaseOPDSImporter[OPDSImporterSettings]):
@@ -1640,313 +1036,3 @@ class OPDSImporter(BaseOPDSImporter[OPDSImporterSettings]):
         series_name = attr.get("{http://schema.org/}name", None)
         series_position = attr.get("{http://schema.org/}position", None)
         return series_name, series_position
-
-
-class OPDSImportMonitor(CollectionMonitor):
-    """Periodically monitor a Collection's OPDS archive feed and import
-    every title it mentions.
-    """
-
-    SERVICE_NAME = "OPDS Import Monitor"
-
-    # The first time this Monitor is invoked we want to get the
-    # entire OPDS feed.
-    DEFAULT_START_TIME = CollectionMonitor.NEVER
-
-    # The protocol this Monitor works with. Subclasses that
-    # specialize OPDS import should override this.
-    PROTOCOL = OPDSAPI.label()
-
-    def __init__(
-        self,
-        _db: Session,
-        collection: Collection,
-        import_class: type[BaseOPDSImporter[OPDSImporterSettings]],
-        force_reimport: bool = False,
-        **import_class_kwargs: Any,
-    ) -> None:
-        if not collection:
-            raise ValueError(
-                "OPDSImportMonitor can only be run in the context of a Collection."
-            )
-
-        if collection.protocol != self.PROTOCOL:
-            raise ValueError(
-                "Collection %s is configured for protocol %s, not %s."
-                % (collection.name, collection.protocol, self.PROTOCOL)
-            )
-
-        self.force_reimport = force_reimport
-
-        self.importer = import_class(_db, collection=collection, **import_class_kwargs)
-        settings = self.importer.settings
-        self.username = settings.username
-        self.password = settings.password
-        self.feed_url = settings.external_account_id
-
-        self.custom_accept_header = settings.custom_accept_header
-        self._max_retry_count = settings.max_retry_count
-
-        parsed_url = urlparse(self.feed_url)
-        self._feed_base_url = f"{parsed_url.scheme}://{parsed_url.hostname}{(':' + str(parsed_url.port)) if parsed_url.port else ''}/"
-        super().__init__(_db, collection)
-
-    def _get(self, url: str, headers: Mapping[str, str]) -> Response:
-        """Make the sort of HTTP request that's normal for an OPDS feed.
-
-        Long timeout, raise error on anything but 2xx or 3xx.
-        """
-
-        headers = self._update_headers(headers)
-        if not url.startswith("http"):
-            url = urljoin(self._feed_base_url, url)
-        return HTTP.get_with_timeout(
-            url,
-            headers=headers,
-            timeout=120,
-            max_retry_count=self._max_retry_count,
-            allowed_response_codes=["2xx", "3xx"],
-        )
-
-    def _get_accept_header(self) -> str:
-        return ",".join(
-            [
-                OPDSFeed.ACQUISITION_FEED_TYPE,
-                "application/atom+xml;q=0.9",
-                "application/xml;q=0.8",
-                "*/*;q=0.1",
-            ]
-        )
-
-    def _update_headers(self, headers: Mapping[str, str] | None) -> dict[str, str]:
-        headers = dict(headers) if headers else {}
-        if self.username and self.password and not "Authorization" in headers:
-            headers["Authorization"] = "Basic %s" % base64.b64encode(
-                f"{self.username}:{self.password}"
-            )
-
-        if self.custom_accept_header:
-            headers["Accept"] = self.custom_accept_header
-        elif not "Accept" in headers:
-            headers["Accept"] = self._get_accept_header()
-
-        return headers
-
-    def data_source(self, collection: Collection) -> DataSource:
-        """Returns the data source name for the given collection.
-
-        By default, this URL is stored as a setting on the collection, but
-        subclasses may hard-code it.
-        """
-        return collection.data_source
-
-    def feed_contains_new_data(self, feed: bytes | str) -> bool:
-        """Does the given feed contain any entries that haven't been imported
-        yet?
-        """
-        if self.force_reimport:
-            # We don't even need to check. Always treat the feed as
-            # though it contained new data.
-            return True
-
-        # For every item in the last page of the feed, check when that
-        # item was last updated.
-        last_update_dates = self.importer.extract_last_update_dates(feed)
-
-        new_data = False
-        for raw_identifier, remote_updated in last_update_dates:
-            identifier = self.importer.parse_identifier(raw_identifier)
-            if not identifier:
-                # Maybe this is new, maybe not, but we can't associate
-                # the information with an Identifier, so we can't do
-                # anything about it.
-                self.log.info(
-                    f"Ignoring {raw_identifier} because unable to turn into an Identifier."
-                )
-                continue
-
-            if self.identifier_needs_import(identifier, remote_updated):
-                new_data = True
-                break
-        return new_data
-
-    def identifier_needs_import(
-        self, identifier: Identifier | None, last_updated_remote: datetime | None
-    ) -> bool:
-        """Does the remote side have new information about this Identifier?
-
-        :param identifier: An Identifier.
-        :param last_update_remote: The last time the remote side updated
-            the OPDS entry for this Identifier.
-        """
-        if not identifier:
-            return False
-
-        record = CoverageRecord.lookup(
-            identifier,
-            self.importer.data_source,
-            operation=CoverageRecord.IMPORT_OPERATION,
-            collection=self.collection,
-        )
-
-        if not record:
-            # We have no record of importing this Identifier. Import
-            # it now.
-            self.log.info(
-                "Counting %s as new because it has no CoverageRecord.", identifier
-            )
-            return True
-
-        # If there was a transient failure last time we tried to
-        # import this book, try again regardless of whether the
-        # feed has changed.
-        if record.status == CoverageRecord.TRANSIENT_FAILURE:
-            self.log.info(
-                "Counting %s as new because previous attempt resulted in transient failure: %s",
-                identifier,
-                record.exception,
-            )
-            return True
-
-        # If our last attempt was a success or a persistent
-        # failure, we only want to import again if something
-        # changed since then.
-
-        if record.timestamp:
-            # We've imported this entry before, so don't import it
-            # again unless it's changed.
-
-            if not last_updated_remote:
-                # The remote isn't telling us whether the entry
-                # has been updated. Import it again to be safe.
-                self.log.info(
-                    "Counting %s as new because remote has no information about when it was updated.",
-                    identifier,
-                )
-                return True
-
-            if to_utc(last_updated_remote) >= to_utc(record.timestamp):
-                # This book has been updated.
-                self.log.info(
-                    "Counting %s as new because its coverage date is %s and remote has %s.",
-                    identifier,
-                    record.timestamp,
-                    last_updated_remote,
-                )
-                return True
-        return False
-
-    def _verify_media_type(self, url: str, resp: Response) -> None:
-        # Make sure we got an OPDS feed, and not an error page that was
-        # sent with a 200 status code.
-        media_type = resp.headers.get("content-type")
-        if not media_type or not any(
-            x in media_type for x in (OPDSFeed.ATOM_LIKE_TYPES)
-        ):
-            message = "Expected Atom feed, got %s" % media_type
-            raise BadResponseException(url, message=message, response=resp)
-
-    def follow_one_link(
-        self, url: str, do_get: Callable[..., Response] | None = None
-    ) -> tuple[list[str], bytes | None]:
-        """Download a representation of a URL and extract the useful
-        information.
-
-        :return: A 2-tuple (next_links, feed). `next_links` is a list of
-            additional links that need to be followed. `feed` is the content
-            that needs to be imported.
-        """
-        self.log.info("Following next link: %s", url)
-        get = do_get or self._get
-        resp = get(url, headers={})
-        feed = resp.content
-
-        self._verify_media_type(url, resp)
-
-        new_data = self.feed_contains_new_data(feed)
-
-        if new_data:
-            # There's something new on this page, so we need to check
-            # the next page as well.
-            next_links = self.importer.extract_next_links(feed)
-            return next_links, feed
-        else:
-            # There's nothing new, so we don't need to import this
-            # feed or check the next page.
-            self.log.info("No new data.")
-            return [], None
-
-    def import_one_feed(
-        self, feed: bytes | str
-    ) -> tuple[list[Edition], dict[str, list[CoverageFailure]]]:
-        """Import every book mentioned in an OPDS feed."""
-
-        # Because we are importing into a Collection, we will immediately
-        # mark a book as presentation-ready if possible.
-        imported_editions, pools, works, failures = self.importer.import_from_feed(
-            feed, feed_url=self.feed_url
-        )
-
-        # Create CoverageRecords for the successful imports.
-        for edition in imported_editions:
-            record = CoverageRecord.add_for(
-                edition,
-                self.importer.data_source,
-                CoverageRecord.IMPORT_OPERATION,
-                status=CoverageRecord.SUCCESS,
-                collection=self.collection,
-            )
-
-        # Create CoverageRecords for the failures.
-        for urn, failure_items in list(failures.items()):
-            for failure_item in failure_items:
-                failure_item.to_coverage_record(
-                    operation=CoverageRecord.IMPORT_OPERATION
-                )
-
-        return imported_editions, failures
-
-    def _get_feeds(self) -> Iterable[tuple[str, bytes]]:
-        feeds = []
-        queue = [self.feed_url]
-        seen_links = set()
-
-        # First, follow the feed's next links until we reach a page with
-        # nothing new. If any link raises an exception, nothing will be imported.
-
-        while queue:
-            new_queue = []
-
-            for link in queue:
-                if link in seen_links:
-                    continue
-                next_links, feed = self.follow_one_link(link)
-                new_queue.extend(next_links)
-                if feed:
-                    feeds.append((link, feed))
-                seen_links.add(link)
-
-            queue = new_queue
-
-        # Start importing at the end. If something fails, it will be easier to
-        # pick up where we left off.
-        return reversed(feeds)
-
-    def run_once(self, progress: TimestampData) -> TimestampData:
-        feeds = self._get_feeds()
-        total_imported = 0
-        total_failures = 0
-
-        for link, feed in feeds:
-            self.log.info("Importing next feed: %s", link)
-            imported_editions, failures = self.import_one_feed(feed)
-            total_imported += len(imported_editions)
-            total_failures += len(failures)
-            self._db.commit()
-
-        achievements = "Items imported: %d. Failures: %d." % (
-            total_imported,
-            total_failures,
-        )
-
-        return TimestampData(achievements=achievements)
