+++ tests/fixtures/database.py
@@ -56,7 +56,8 @@ from palace.manager.integration.license.boundless.api import BoundlessApi
 from palace.manager.integration.license.boundless.settings import BoundlessSettings
 from palace.manager.integration.license.opds.odl.api import OPDS2WithODLApi
 from palace.manager.integration.license.opds.odl.settings import OPDS2WithODLSettings
-from palace.manager.integration.license.opds.opds1 import OPDSAPI, OPDSImporterSettings
+from palace.manager.integration.license.opds.opds1.api import OPDSAPI
+from palace.manager.integration.license.opds.opds1.settings import OPDSImporterSettings
 from palace.manager.integration.license.opds.opds2.api import OPDS2API
 from palace.manager.integration.license.opds.opds_for_distributors import (
     OPDSForDistributorsAPI,
