+++ tests/manager/api/controller/test_multilib.py
@@ -1,4 +1,4 @@
-from palace.manager.integration.license.opds.opds1 import OPDSAPI
+from palace.manager.integration.license.opds.opds1.api import OPDSAPI
 from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.patron import Patron
 from tests.fixtures.api_controller import (
