+++ tests/manager/integration/license/opds/test_opds_for_distributors.py
@@ -17,7 +17,7 @@ from palace.manager.api.circulation.fulfillment import (
 from palace.manager.data_layer.circulation import CirculationData
 from palace.manager.data_layer.identifier import IdentifierData
 from palace.manager.data_layer.link import LinkData
-from palace.manager.integration.license.opds import opds1
+from palace.manager.integration.license.opds.base import importer
 from palace.manager.integration.license.opds.opds_for_distributors import (
     OPDSForDistributorsAPI,
     OPDSForDistributorsImporter,
@@ -787,7 +787,7 @@ class TestOPDSForDistributorsImporter:
             collection=collection2,
         )
 
-        with patch.object(opds1, "get_one", wraps=get_one) as get_one_mock:
+        with patch.object(importer, "get_one", wraps=get_one) as get_one_mock:
             importer1_lp, _ = importer1.update_work_for_edition(edition)
             importer2_lp, _ = importer2.update_work_for_edition(edition)
