+++ tests/manager/scripts/test_opds_import.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from palace.manager.integration.license.opds.opds1 import OPDSImportMonitor
+from palace.manager.integration.license.opds.opds1.monitor import OPDSImportMonitor
 from palace.manager.scripts.opds_import import OPDSImportScript
 from tests.fixtures.database import DatabaseTransactionFixture
